myaddonsext/ygsoft/requirements.txt
python_ldap
pysmb
pyasn1